# lib/api-zod/src/generated/types/openaiConversationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Platform API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiConversationInput {
  title: string;
}